requirements.txt
Flask
gunicorn
flask-dance
google-generativeai
